Linked list: Added callback for all remove methods
Not sure whether this will be good or just a horrible
design action.  Now, all remove methods, which may
delete a node from the linked list, have a callback
parameter to free a node.  Probably will be the same
callback for all methods, but to add a property
for this would be expensive, in terms of refactoring.
Signed-off-by: Paulo Roberto Urio <paulourio@gmail.com>